requirements.txt
# Core dependencies
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
torch>=2.2.0
scikit-learn>=1.4.0
wandb>=0.17.0
albumentations>=1.4.0
